kdtree is no longer enforced
No longer enforces the knnsearch nearest neighbor search method to
'kdtree'. It is better to not set it and let MATLAB choose based on its
own criteria. It basically chooses 'kdtree' when the number of features
is 10 or less and 'exhaustive' otherwise. For some distance functions,
including 'seuclidean', 'exhaustive' is the only option.